UI Error is Fixed and the Level Restarts when needed to
Hello Gamers, it's Preston Alexander Cherry here to fix an error with the UI being unable to do work. The Fix was very complicated, adding a Event System to each scene that requires the UI Pause menu and such. Plus the Level Select works fine when the level is completed without fail. : ) XD